ISSUE #320 update header

+++ bouncer/src/repo/core/model/bson/repo_bson_factory.h
@@ -95,7 +95,7 @@ namespace repo {
 					const std::string &fileName,
 					const RepoRef::RefType &type,
 					const std::string &link,
-					const size_t size);
+					const uint32_t size);
 
 				/**
 				* Create a role BSON
